moved constants data
